DVD-476 add java exception when file doesn't exist or is a directory and code refactoring
